+++ common/src/main/kotlin/com/cobblemon/mod/common/entity/pokemon/PokemonEntity.kt
@@ -58,7 +58,7 @@ import com.cobblemon.mod.common.pokemon.activestate.InactivePokemonState
 import com.cobblemon.mod.common.pokemon.activestate.ShoulderedState
 import com.cobblemon.mod.common.pokemon.ai.FormPokemonBehaviour
 import com.cobblemon.mod.common.pokemon.evolution.variants.ItemInteractionEvolution
-import com.cobblemon.mod.common.pokemon.misc.GimmighoulStashHandler
+import com.cobblemon.mod.common.pokemon.feature.StashHandler
 import com.cobblemon.mod.common.util.*
 import com.cobblemon.mod.common.world.gamerules.CobblemonGameRules
 import java.util.EnumSet
@@ -710,16 +710,8 @@ open class PokemonEntity(
                 pokemon.lastFlowerFed = itemStack
                 return ActionResult.success(world.isClient)
             }
-        } else if(!player.isSneaking && (itemStack.isOf(CobblemonItems.RELIC_COIN)
-                || itemStack.isOf(CobblemonItems.RELIC_COIN_POUCH)
-                || itemStack.isOf(CobblemonItems.RELIC_COIN_SACK)
-                || itemStack.isOf(Items.NETHERITE_SCRAP)
-                || itemStack.isOf(Items.NETHERITE_INGOT)
-                || itemStack.isOf(Items.NETHERITE_BLOCK))) {
-
-            if(GimmighoulStashHandler.interactMob(player, hand, pokemon)) {
-                return ActionResult.SUCCESS
-            }
+        } else if(!player.isSneaking && StashHandler.interactMob(player, pokemon, itemStack)) {
+            return ActionResult.SUCCESS
         }
         else if (itemStack.item is DyeItem && colorFeature != null) {
             val item = itemStack.item as DyeItem